requirements.txt

streamlit
streamlit_option_menu
pandas
numpy
scikit-learn
nltk
pyyaml